+++ passwordvault_api/passwordvault_domain/Repositories/IQueryRepository.cs
@@ -2,5 +2,5 @@ namespace passwordvault_domain.Repositories;
 
 public interface IQueryRepository<T>
 {
-    Task<T> GetById(int id);
+    Task<T> GetById(long id);
 }
